I/O: FreeImageIO seeks and tells with 64-bit positions
FI_SeekProc took a long offset and FI_TellProc returned one, and long is
32 bits on Windows: nothing past 2 GB of a file could be addressed there,
through FreeImage_Load() and FreeImage_Save() as much as through a
FreeImageIO of the caller's. Three plugins (JXR, HEIF, AVIF) carried their
own position and stepped past LONG_MAX with SEEK_CUR, PSD skipped its layer
section in 256 MB steps, and every other plugin kept the positions it
needed in a long. Both callbacks take and return INT64 now; the layout of
FreeImageIO and every export are unchanged, so a C caller recompiles (a
FreeImageIO implemented with long is refused at compile time), and only
old Windows binaries with callbacks of their own would break: QPV and the
wrappers implement none.

- the default I/O uses _fseeki64/_ftelli64 on Windows and fseeko/ftello
  elsewhere (FreeImage_fseek64/ftell64, which the multi-page cache file and
  the in-place JPEG transform use too: the cache seeked with a 32-bit
  product past 2 GB, the transform truncated with _chsize);
- memory streams are INT64-sized: FreeImage_OpenMemory() with a buffer of
  2 GB or more read nothing (its length went negative), and a stream
  FreeImage grows stopped at 2 GB; both now hold what memory allows, while
  the exports keep their types: FreeImage_AcquireMemory() returns FALSE past
  4 GB and FreeImage_TellMemory() -1 past what a long holds;
- read_proc/write_proc keep their unsigned counts; FreeImage_ReadBytes() and
  FreeImage_WriteBytes() move any size through them in 1 GB pieces, and the
  glue of libtiff, jxrlib, OpenJPEG, LibRaw, libheif, libavif, libwebp, G3
  and APNG uses them instead of an (unsigned) cast that silently shortened a
  read of 4 GB or more;
- every position a plugin holds is INT64 (62 tell_proc results, the
  offset/size parameters of MNG, PNG, TARGA, EXR and BMP, GIF's size_t seeks,
  the stepped seeks of JXR, HEIF, AVIF, PSD and RAS are single seeks now);
  G3 also reads a stream that does not start at byte zero;
- the wrappers follow: Delphi Int64, .NET long, the C++/CLI shim INT64; the
  AHK wrapper only documents the change (it implements no callbacks);
- TestAPI/IO: sparse BigTIFFs with pages at 2.5 GB and 4.5 GB read through
  every entry point, a 2.5 GB wrapped memory stream, a stream grown past
  2 GB, and a 2.2 GB TIFF save; the JXR, HEIF and AVIF narrowio tests keep
  their handle and offset-stream checks and drop the emulated 32-bit tells
  and the narrowio-step builds, which tested code that no longer exists.

Checked: full Linux build, testAPI and the IO, JXR, HEIF, AVIF, J2K, EXR,
MNG, APNG, ICC, WebP and RAW suites pass; the whole library builds for
Windows x64 with zig's clang, and under Wine the IO tests pass against it
where the same binaries built against the old header and library fail past
2 GB; a -Wshorten-64-to-32 diff of the changed sources for Win64 shows no
new narrowing.

+++ Source/FreeImage/MultiPage.cpp
@@ -152,7 +152,7 @@ struct MULTIBITMAPHEADER {
 	// plugin decoder state, from the first LockPage until close
 	void *read_data;
 	// handle position where the image starts; a stream need not hold it at 0
-	long start;
+	INT64 start;
 };
 
 // =====================================================================
@@ -599,7 +599,7 @@ FreeImage_OpenMultiBitmapFromHandle(FREE_IMAGE_FORMAT fif, FreeImageIO *io, fi_h
 					header->node = node;
 					header->fif = fif;
 					header->handle = handle;						
-					header->start = MAX(io->tell_proc(handle), 0L);
+					header->start = MAX(io->tell_proc(handle), (INT64)0);
 					header->read_only = read_only;	
 					header->cache_fif = fif;
 					header->load_flags = flags;
